requirements.txt
trio-websocket==0.12.2
aiofiles==25.1.0
asyncclick==8.1.8.0
pytest-asyncio==1.2.0
pytest-trio==0.8.0
